Breaking up runner.c to mirror haskell version of the runner
