+++ rts-pgen-c/runner.h
@@ -5,11 +5,6 @@
 
 #include "aut.h"
 
-typedef struct _Input {
-    FILE * file;
-    fpos_t posInput;
-} Input ;
-
 extern void runAut(Aut aut, FILE* input);
 
 #endif /* RUNNER_H */
